requirements.txt
torch==2.0.0+cu117
Pillow==9.4.0
transformers==4.47.0
numpy==1.24.2